trinity_engine/src/requirements.txt
customtkinter
psutil
humanize
wmi
Pillow
pywin32
